trip-review-summary/pyproject.toml
[project]
name = "review-summary"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server,telemetry]>=0.3.22",
    "asgiref>=3.11.0",
    "celery[redis]>=5.6.1",
    "fastapi[standard]>=0.128.0",
    "graphdatascience>=1.18",
    "grpcio>=1.76.0",
    "langchain-openai>=1.1.6",
    "nacos-sdk-python>=3.0.3",
    "networkx>=3.6.1",
    "openinference-instrumentation-langchain>=0.1.58",
    "opentelemetry-distro[otlp]>=0.60b1",
    "pydantic-settings>=2.12.0",
    "qdrant-client>=1.16.2",
    "tripsphere>=0.1.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "celery-types>=0.24.0",
    "mypy>=1.19.1",
    "pandas-stubs>=2.3.3.251219",
    "pyarrow-stubs>=20.0.0.20251215",
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "types-grpcio>=1.0.0.20251009",
    "types-networkx>=3.6.1.20251220",
    "types-protobuf>=6.32.1.20251210",
]

[tool.uv.sources]
tripsphere = { workspace = true }

[tool.uv.workspace]
members = ["libs/tripsphere"]

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]

[tool.pytest]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
